# specs/001-weatherstation-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WeatherStation Backend API (Red de Estaciones)
  description: >
    API REST del backend central de la Red de Estaciones Meteorológicas
    Inteligentes (CLIMBOT). Único componente con acceso a la base de datos
    PostgreSQL (almacén único: sistema + lecturas meteorológicas) y a Gemini (IA).
    Las estaciones ESP32 publican sus lecturas a través de `/api/device/**` (nunca
    a un almacén externo). Autenticación JWT (Bearer) con roles ADMIN, RESPONSABLE,
    INVESTIGADOR y USUARIO para usuarios, y token de estación → JWT de dispositivo
    para los ESP32.
  version: 3.1.0
  contact:
    name: WeatherStation
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://weatherstation-backend.onrender.com
    description: Producción (Render)
...
security:
  - bearerAuth: []
paths:
...
  /estaciones:
    get:
      tags: [Estaciones]
      summary: Listar estaciones (visibilidad según rol)
      description: >
        USUARIO/INVESTIGADOR ven solo APPROVED+habilitadas; RESPONSABLE además las
        de su escuela en cualquier estado; ADMIN todas. Filtros opcionales.
      parameters:
        - { in: query, name: escuelaId, schema: { type: string, format: uuid } }
        - { in: query, name: estado, schema: { $ref: '#/components/schemas/EstadoEstacion' } }
      responses:
        '200':
          description: Lista de estaciones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EstacionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Estaciones]
      summary: Registrar una estación (ADMIN o RESPONSABLE de la escuela) → PENDING
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionRegistroRequest' }
      responses:
        '201':
          description: Estación creada en estado PENDING
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /estaciones/{id}:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    get:
      tags: [Estaciones]
      summary: Detalle de una estación
      responses:
        '200':
          description: Estación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Estaciones]
      summary: Actualizar metadata de la estación (ADMIN o RESPONSABLE de su escuela)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionUpdateRequest' }
      responses:
        '200':
          description: Estación actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Estaciones]
      summary: Eliminar una estación (ADMIN)
      responses:
        '204': { description: Eliminada }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/aprobar:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Aprobar una estación (ADMIN) → APPROVED + genera token (una vez)
      responses:
        '200':
          description: Estación aprobada; token devuelto una única vez
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StationTokenResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La estación no está en un estado aprobable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /estaciones/{id}/rechazar:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Rechazar una estación (ADMIN) → REJECTED
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionAccionRequest' }
      responses:
        '200':
          description: Estación rechazada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/deshabilitar:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Deshabilitar una estación aprobada (ADMIN) → DISABLED
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionAccionRequest' }
      responses:
        '200':
          description: Estación deshabilitada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/reactivar:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Reactivar una estación deshabilitada/en mantenimiento (ADMIN) → APPROVED
      responses:
        '200':
          description: Estación reactivada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/mantenimiento:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Poner una estación en mantenimiento (ADMIN) → MAINTENANCE
      responses:
        '200':
          description: Estación en mantenimiento (no podrá publicar)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/regenerar-token:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    post:
      tags: [Estaciones]
      summary: Regenerar el token de la estación (ADMIN) — revoca el anterior
      responses:
        '200':
          description: Token regenerado; devuelto una única vez
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StationTokenResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/conexiones:
    parameters:
      - { $ref: '#/components/parameters/EstacionId' }
      - { in: query, name: page, schema: { type: integer, default: 0 } }
      - { in: query, name: size, schema: { type: integer, default: 50 } }
    get:
      tags: [Estaciones]
      summary: Historial de conexiones de la estación (ADMIN o RESPONSABLE)
      responses:
        '200':
          description: Conexiones registradas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConexionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────── ESTACIONES (datos meteorológicos, lectura) ──────────
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar usuarios (ADMIN)
      parameters:
        - { in: query, name: page, schema: { type: integer, default: 0 } }
        - { in: query, name: size, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Lista paginada de usuarios
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageUsuarioResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Usuarios]
      summary: Crear usuario con rol (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuarioRequest' }
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /usuarios/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Usuarios]
      summary: Ver un usuario (ADMIN)
      responses:
        '200':
          description: Usuario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Usuarios]
      summary: Actualizar un usuario (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuarioUpdateRequest' }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuarioResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Usuarios]
      summary: Eliminar un usuario (ADMIN)
      responses:
        '204': { description: Eliminado }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────────── IA ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de usuario (claim `rol`).
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de dispositivo emitido por /api/device/auth (claim `type=DEVICE`).
  parameters:
    EstacionId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: UUID público de la estación
      example: 7b1f2c34-5d6e-4a8b-9c0d-1e2f3a4b5c6d
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: No autenticado o credenciales inválidas
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Rol/recurso no autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto (p. ej. email/clave ya existe)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    Rol:
      type: string
      enum: [ADMIN, RESPONSABLE, INVESTIGADOR, USUARIO]
    EstadoEstacion:
      type: string
      enum: [PENDING, APPROVED, REJECTED, DISABLED, MAINTENANCE]
    Conectividad:
      type: string
      enum: [ONLINE, OFFLINE]
      description: Derivado de ultimaConexion + umbral; no es un estado almacenado.
...
    UsuarioRequest:
      type: object
      required: [nombre, email, password, rol]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72, format: password }
        rol: { $ref: '#/components/schemas/Rol' }
        escuelaId:
          type: string
          format: uuid
          description: Requerido si rol = RESPONSABLE
    UsuarioUpdateRequest:
      type: object
      properties:
        nombre: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        rol: { $ref: '#/components/schemas/Rol' }
        escuelaId: { type: string, format: uuid }
        activo: { type: boolean }
        password: { type: string, minLength: 8, maxLength: 72, format: password }
    UsuarioResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        email: { type: string, format: email }
        rol: { $ref: '#/components/schemas/Rol' }
        escuelaId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        escuelaNombre: { oneOf: [ { type: string }, { type: 'null' } ] }
        activo: { type: boolean }
        createdAt: { type: string, format: date-time }
    PageUsuarioResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/UsuarioResponse' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    EstacionRegistroRequest:
      type: object
      required: [nombre, escuelaId]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 120 }
        escuelaId: { type: string, format: uuid }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        firmware: { type: string, maxLength: 40 }
    EstacionUpdateRequest:
      type: object
      properties:
        nombre: { type: string, maxLength: 120 }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        responsableId: { type: string, format: uuid }
    EstacionAccionRequest:
      type: object
      properties:
        motivo: { type: string, maxLength: 300 }
    EstacionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        uuid: { type: string, format: uuid }
        nombre: { type: string }
        descripcion: { type: string }
        escuelaId: { type: string, format: uuid }
        escuelaNombre: { type: string }
        responsableId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        responsableNombre: { oneOf: [ { type: string }, { type: 'null' } ] }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        altitud: { type: number, format: double }
        firmware: { type: string }
        hardware: { type: string }
        ultimoRssi: { oneOf: [ { type: integer }, { type: 'null' } ] }
        estado: { $ref: '#/components/schemas/EstadoEstacion' }
        conectividad: { $ref: '#/components/schemas/Conectividad' }
        fechaRegistro: { type: string, format: date-time }
        ultimaConexion: { oneOf: [ { type: string, format: date-time }, { type: 'null' } ] }
        enLinea: { type: boolean }
        ultimaLectura:
          oneOf:
            - { $ref: '#/components/schemas/LecturaResponse' }
            - { type: 'null' }
    StationTokenResponse:
      type: object
      description: El token en claro se muestra una única vez y no se vuelve a exponer.
      properties:
        estacionId: { type: string, format: uuid }
        uuid: { type: string, format: uuid }
        token: { type: string, description: Token permanente de la estación (en claro, solo aquí) }
        generadoEn: { type: string, format: date-time }
        aviso: { type: string, example: "Guarde este token: no se volverá a mostrar." }
    ConexionResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        ip: { type: string }
        firmware: { type: string }
        evento: { type: string, example: DATA_OK }
        detalle: { type: string }
...
    LecturaResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        temperatura: { type: number, format: double, example: 24.5 }
        humedad: { type: number, format: double, example: 60 }
        presion: { type: number, format: double, example: 1013.2 }
        vientoKmh: { type: number, format: double, example: 12.3 }
        vientoDir: { type: string, example: NE }
        vientoGrados:
          oneOf:
            - { type: number, format: double }
            - { type: 'null' }
          description: Solo presente en la lectura actual; null en historial
        lluviaMm: { type: number, format: double, example: 0.0 }
...
    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        code: { type: string, example: ESTACION_NO_APROBADA }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }